AddNewTaskBtn updates
